modified new_entry_r.py, moved date button here and rearanged view
